move heavy test to e2e
