+++ tests/utils/test_models.py
@@ -3,14 +3,13 @@
 from unittest.mock import MagicMock, patch
 
 import pytest
-import torch
 from transformers import BitsAndBytesConfig, PreTrainedTokenizerBase
 from transformers.integrations.deepspeed import is_deepspeed_zero3_enabled
 from transformers.utils.import_utils import is_torch_mps_available
 
 from axolotl.cli import load_cfg
 from axolotl.utils.dict import DictDefault
-from axolotl.utils.models import ModelLoader, load_model, load_tokenizer
+from axolotl.utils.models import ModelLoader, load_model
 
 
 class TestModelsUtils:
@@ -112,29 +111,3 @@ class TestModelsUtils:
             assert self.model_loader.model_kwargs.get(
                 "quantization_config", BitsAndBytesConfig
             )
-
-    @pytest.mark.parametrize("embedding_modules", ["embed_tokens", "lm_head"])
-    @pytest.mark.parametrize(
-        "dist_dtype", [torch.bfloat16, torch.float16, torch.float32]
-    )
-    @pytest.mark.parametrize("before_kbit_train_or_finetune", [True, False])
-    def test_convert_embedding_modules_dtype(
-        self, embedding_modules, dist_dtype, before_kbit_train_or_finetune
-    ):
-        tokenizer = load_tokenizer(self.cfg)
-        self.model_loader.model, _ = load_model(self.cfg, tokenizer, inference=False)
-
-        self.model_loader.convert_embedding_modules_dtype(
-            embedding_modules, dist_dtype, before_kbit_train_or_finetune
-        )
-        for name, module in self.model_loader.model.named_modules():
-            if (
-                "norm" in name
-                or (before_kbit_train_or_finetune and name.endswith(".gate"))
-                or (
-                    any(m in name for m in embedding_modules)
-                    and hasattr(module, "weight")
-                )
-            ):
-                for _, param in module.named_parameters(recurse=False):
-                    assert param.dtype == dist_dtype
